+++ src/Admin.Api/Models/Requests.cs
@@ -2,27 +2,27 @@ namespace ScalaAPI.Admin.Models;
 
 public record AccountCreateRequest(
     string Name, string Platform, string Type, string BaseUrl,
-    int Priority, int Concurrency, int LoadFactor, double RateMultiplier,
+    int Priority, int Concurrency, int LoadFactor, decimal RateMultiplier,
     bool Schedulable, Dictionary<string, string> Credentials,
     Dictionary<string, string> ModelMapping, string[] SupportedModels,
     string? ProxyUrl, bool TlsFingerprint);
 
 public record GroupCreateRequest(
-    string Platform, double RateMultiplier, bool IsExclusive,
-    double? DailyLimitUsd, bool ClaudeCodeOnly, long? FallbackGroupId,
+    string Platform, decimal RateMultiplier, bool IsExclusive,
+    decimal? DailyLimitUsd, bool ClaudeCodeOnly, long? FallbackGroupId,
     bool ModelRoutingEnabled, Dictionary<string, long[]> ModelRouting,
     long[] MemberAccountIds, int RpmLimit,
-    double? PeakMultiplier, int? PeakStartHour, int? PeakEndHour);
+    decimal? PeakMultiplier, int? PeakStartHour, int? PeakEndHour);
 
 public record UserCreateRequest(
-    string Role, double Balance, int Concurrency,
+    string Role, decimal Balance, int Concurrency,
     int RpmLimit, long[] AllowedGroups);
 
 public record ApiKeyCreateRequest(
-    long UserId, long GroupId, double Quota,
+    long UserId, long GroupId, decimal Quota,
     long? ExpiresAt, string[] IpWhitelist, string[] IpBlacklist,
-    double RateLimit5h, double RateLimit1d, double RateLimit7d);
+    decimal RateLimit5h, decimal RateLimit1d, decimal RateLimit7d);
 
 public record StatusRequest(string Status);
-public record BalanceRequest(double Delta);
+public record BalanceRequest(decimal Delta);
 public record ConfigUpdateRequest(string Key, string Value);
